Make gardenlet's `NetworkPolicy` controller reusable
