+++ pkg/controller/networkpolicy/peers.go
